Warning: Untested!
Restructured project folders
Separated generic chunks into their own header
Changed use of ints to longs
Added and switching to macro's for differing variable types
Create/Copy/Delete GMST functions removed, and their functionality added to unified functions
More commented out code removed
Collection::RemoveIndex removed
Collection::DeleteCELL removed
Record visitation improved, VisitTopRecords removed
RecordDeleter visitor replaces the various RecordDelete functions
STRING class renamed to StringRecord
ISTRING class renamed to InsensitiveStringRecord
RAWBYTES class renamed to RawRecord
BaseRecord.h cleaned up and code split into .cpp
Collection::LookupGMSTRecord renamed LookupRecordByEditorID and works with any record keyed by editorID
Fixed API access to AMMORecord::DATA::flags, it's a UInt16, but was accessed like it was a UInt32.
_FormIDHandler class renamed to FormIDHandlerClass

+++ CBash/Records/ACHRRecord.cpp
@@ -22,14 +22,14 @@ GPL License and Copyright Notice ============================================
 #include "..\Common.h"
 #include "ACHRRecord.h"
 
-int ACHRRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
+signed long ACHRRecord::ParseRecord(unsigned char *buffer, const unsigned long &recSize)
     {
     if(IsLoaded())
         return -1;
     IsLoaded(true);
-    unsigned int subType = 0;
-    unsigned int subSize = 0;
-    unsigned int curPos = 0;
+    unsigned long subType = 0;
+    unsigned long subSize = 0;
+    unsigned long curPos = 0;
     while(curPos < recSize){
         _readBuffer(&subType,buffer,4,curPos);
         switch(subType)
@@ -94,12 +94,12 @@ int ACHRRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
     return 0;
     }
 
-unsigned int ACHRRecord::GetSize(bool forceCalc)
+unsigned long ACHRRecord::GetSize(bool forceCalc)
     {
     if(!forceCalc && recData != NULL)
-        return *(unsigned int*)&recData[-16];
-    unsigned int cSize = 0;
-    unsigned int TotSize = 0;
+        return *(unsigned long*)&recData[-16];
+    unsigned long cSize = 0;
+    unsigned long TotSize = 0;
 
     if(EDID.IsLoaded())
         {
@@ -151,7 +151,7 @@ unsigned int ACHRRecord::GetSize(bool forceCalc)
     return TotSize;
     }
 
-int ACHRRecord::WriteRecord(_FileHandler &SaveHandler)
+signed long ACHRRecord::WriteRecord(_FileHandler &SaveHandler)
     {
     char null = 0;
 
@@ -198,7 +198,7 @@ void ACHRRecord::Debug(int debugLevel)
     {
     if(!IsLoaded())
         return;
-    unsigned int indentation = 4;
+    unsigned long indentation = 4;
     printf("  ACHR\n");
     if(Header.IsLoaded())
         Header.Debug(debugLevel, indentation);
