+++ CBash/Records/BOOKRecord.cpp
@@ -22,14 +22,14 @@ GPL License and Copyright Notice ============================================
 #include "..\Common.h"
 #include "BOOKRecord.h"
 
-int BOOKRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
+signed long BOOKRecord::ParseRecord(unsigned char *buffer, const unsigned long &recSize)
     {
     if(IsLoaded())
         return -1;
     IsLoaded(true);
-    unsigned int subType = 0;
-    unsigned int subSize = 0;
-    unsigned int curPos = 0;
+    unsigned long subType = 0;
+    unsigned long subSize = 0;
+    unsigned long curPos = 0;
     while(curPos < recSize){
         _readBuffer(&subType,buffer,4,curPos);
         switch(subType)
@@ -95,12 +95,12 @@ int BOOKRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
     return 0;
     }
 
-unsigned int BOOKRecord::GetSize(bool forceCalc)
+unsigned long BOOKRecord::GetSize(bool forceCalc)
     {
     if(!forceCalc && recData != NULL)
-        return *(unsigned int*)&recData[-16];
-    unsigned int cSize = 0;
-    unsigned int TotSize = 0;
+        return *(unsigned long*)&recData[-16];
+    unsigned long cSize = 0;
+    unsigned long TotSize = 0;
     if(EDID.IsLoaded())
         {
         cSize = EDID.GetSize();
@@ -150,7 +150,7 @@ unsigned int BOOKRecord::GetSize(bool forceCalc)
     return TotSize;
     }
 
-int BOOKRecord::WriteRecord(_FileHandler &SaveHandler)
+signed long BOOKRecord::WriteRecord(_FileHandler &SaveHandler)
     {
     if(EDID.IsLoaded())
         SaveHandler.writeSubRecord(eEDID, EDID.value, EDID.GetSize());
@@ -184,7 +184,7 @@ void BOOKRecord::Debug(int debugLevel)
     {
     if(!IsLoaded())
         return;
-    unsigned int indentation = 4;
+    unsigned long indentation = 4;
     printf("  BOOK\n");
     if(Header.IsLoaded())
         Header.Debug(debugLevel, indentation);
