+++ CBash/Records/CLMTRecord.cpp
@@ -23,14 +23,14 @@ GPL License and Copyright Notice ============================================
 #include "CLMTRecord.h"
 #include <vector>
 
-int CLMTRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
+signed long CLMTRecord::ParseRecord(unsigned char *buffer, const unsigned long &recSize)
     {
     if(IsLoaded())
         return -1;
     IsLoaded(true);
-    unsigned int subType = 0;
-    unsigned int subSize = 0;
-    unsigned int curPos = 0;
+    unsigned long subType = 0;
+    unsigned long subSize = 0;
+    unsigned long curPos = 0;
     while(curPos < recSize){
         _readBuffer(&subType,buffer,4,curPos);
         switch(subType)
@@ -98,12 +98,12 @@ int CLMTRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
     return 0;
     }
 
-unsigned int CLMTRecord::GetSize(bool forceCalc)
+unsigned long CLMTRecord::GetSize(bool forceCalc)
     {
     if(!forceCalc && recData != NULL)
-        return *(unsigned int*)&recData[-16];
-    unsigned int cSize = 0;
-    unsigned int TotSize = 0;
+        return *(unsigned long*)&recData[-16];
+    unsigned long cSize = 0;
+    unsigned long TotSize = 0;
     if(EDID.IsLoaded())
         {
         cSize = EDID.GetSize();
@@ -112,7 +112,7 @@ unsigned int CLMTRecord::GetSize(bool forceCalc)
         }
     if(Weathers.size())
         {
-        cSize = (sizeof(CLMTWLST) * (unsigned int)Weathers.size());
+        cSize = (sizeof(CLMTWLST) * (unsigned long)Weathers.size());
         if(cSize > 65535) cSize += 10;
         TotSize += cSize += 6;
         }
@@ -149,12 +149,12 @@ unsigned int CLMTRecord::GetSize(bool forceCalc)
     return TotSize;
     }
 
-int CLMTRecord::WriteRecord(_FileHandler &SaveHandler)
+signed long CLMTRecord::WriteRecord(_FileHandler &SaveHandler)
     {
     if(EDID.IsLoaded())
         SaveHandler.writeSubRecord(eEDID, EDID.value, EDID.GetSize());
     if(Weathers.size())
-        SaveHandler.writeSubRecord(eWLST, &Weathers[0], (unsigned int)Weathers.size() * sizeof(CLMTWLST));
+        SaveHandler.writeSubRecord(eWLST, &Weathers[0], (unsigned long)Weathers.size() * sizeof(CLMTWLST));
     //else
     //    SaveHandler.writeSubRecord(eWLST, NULL, 0);
 
@@ -182,7 +182,7 @@ void CLMTRecord::Debug(int debugLevel)
     {
     if(!IsLoaded())
         return;
-    unsigned int indentation = 4;
+    unsigned long indentation = 4;
     printf("  CLMT\n");
     if(Header.IsLoaded())
         Header.Debug(debugLevel, indentation);
@@ -194,7 +194,7 @@ void CLMTRecord::Debug(int debugLevel)
         PrintIndent(indentation);
         printf("Weathers:\n");
         indentation += 2;
-        for(unsigned int p = 0;p < Weathers.size();p++)
+        for(unsigned long p = 0;p < Weathers.size();p++)
             {
             PrintIndent(indentation);
             printf("Index: %u\n", p);
