+++ CBash/Records/CLOTRecord.cpp
@@ -23,14 +23,14 @@ GPL License and Copyright Notice ============================================
 #include "CLOTRecord.h"
 #include <vector>
 
-int CLOTRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
+signed long CLOTRecord::ParseRecord(unsigned char *buffer, const unsigned long &recSize)
     {
     if(IsLoaded())
         return -1;
     IsLoaded(true);
-    unsigned int subType = 0;
-    unsigned int subSize = 0;
-    unsigned int curPos = 0;
+    unsigned long subType = 0;
+    unsigned long subSize = 0;
+    unsigned long curPos = 0;
     while(curPos < recSize){
         _readBuffer(&subType,buffer,4,curPos);
         switch(subType)
@@ -140,12 +140,12 @@ int CLOTRecord::ParseRecord(unsigned char *buffer, const unsigned int &recSize)
     return 0;
     }
 
-unsigned int CLOTRecord::GetSize(bool forceCalc)
+unsigned long CLOTRecord::GetSize(bool forceCalc)
     {
     if(!forceCalc && recData != NULL)
-        return *(unsigned int*)&recData[-16];
-    unsigned int cSize = 0;
-    unsigned int TotSize = 0;
+        return *(unsigned long*)&recData[-16];
+    unsigned long cSize = 0;
+    unsigned long TotSize = 0;
     if(EDID.IsLoaded())
         {
         cSize = EDID.GetSize();
@@ -239,7 +239,7 @@ unsigned int CLOTRecord::GetSize(bool forceCalc)
     return TotSize;
     }
 
-int CLOTRecord::WriteRecord(_FileHandler &SaveHandler)
+signed long CLOTRecord::WriteRecord(_FileHandler &SaveHandler)
     {
     if(EDID.IsLoaded())
         SaveHandler.writeSubRecord(eEDID, EDID.value, EDID.GetSize());
@@ -301,7 +301,7 @@ void CLOTRecord::Debug(int debugLevel)
     {
     if(!IsLoaded())
         return;
-    unsigned int indentation = 4;
+    unsigned long indentation = 4;
     printf("  CLOT\n");
     if(Header.IsLoaded())
         Header.Debug(debugLevel, indentation);
